[#6] DB-based storage: fixtures' refactoring and coverage configured
